+++ cucumber-expressions/javascript/src/CucumberExpressionGenerator.ts
@@ -1,7 +1,6 @@
 import ParameterTypeMatcher from './ParameterTypeMatcher'
 import ParameterType from './ParameterType'
 
-import util from 'util'
 import CombinatorialGeneratedExpressionFactory from './CombinatorialGeneratedExpressionFactory'
 import GeneratedExpression from './GeneratedExpression'
 
@@ -13,7 +12,7 @@ export default class CucumberExpressionGenerator {
     const parameterTypeMatchers = this.createParameterTypeMatchers(text)
     let expressionTemplate = ''
     let pos = 0
-
+    let counter = 0
     // eslint-disable-next-line no-constant-condition
     while (true) {
       let matchingParameterTypeMatchers = []
@@ -53,7 +52,7 @@ export default class CucumberExpressionGenerator {
         parameterTypeCombinations.push(parameterTypes)
 
         expressionTemplate += escape(text.slice(pos, bestParameterTypeMatcher.start))
-        expressionTemplate += '{%s}'
+        expressionTemplate += `{{${counter++}}}`
 
         pos = bestParameterTypeMatcher.start + bestParameterTypeMatcher.group.length
       } else {
@@ -70,16 +69,6 @@ export default class CucumberExpressionGenerator {
       expressionTemplate,
       parameterTypeCombinations
     ).generateExpressions()
-  }
-
-  /**
-   * @deprecated
-   */
-  public generateExpression(text: string): GeneratedExpression {
-    return util.deprecate(
-      () => this.generateExpressions(text)[0],
-      'CucumberExpressionGenerator.generateExpression: Use CucumberExpressionGenerator.generateExpressions instead'
-    )()
   }
 
   private createParameterTypeMatchers(text: string): ParameterTypeMatcher[] {
@@ -109,10 +98,7 @@ export default class CucumberExpressionGenerator {
 
 function escape(s: string): string {
   return s
-    .replace(/%/g, '%%') // for util.format
     .replace(/\(/g, '\\(')
     .replace(/{/g, '\\{')
     .replace(/\//g, '\\/')
 }
-
-module.exports = CucumberExpressionGenerator