+++ monaco/javascript/example/index.js
@@ -1,17 +1,7 @@
 import * as monaco from 'monaco-editor/esm/vs/editor/editor.api';
 import { buildStepDocuments, jsSearchIndex } from '@cucumber/suggest'
 import { ExpressionFactory, ParameterTypeRegistry } from '@cucumber/cucumber-expressions'
-import { configure } from '../dist/src/index'
-
-(function () {
-  // create div to avoid needing a HtmlWebpackPlugin template
-  const div = document.createElement('div');
-  div.id = 'root';
-  // @ts-ignore
-  div.style = 'width:800px; height:600px; border:1px solid #ccc;';
-
-  document.body.appendChild(div);
-})();
+import { configureMonaco } from '../dist/src/index'
 
 // Build some sample step texts and cucumber expressions. These would typically come from a stream
 // of Cucumber Messages.
@@ -30,17 +20,32 @@ const docs = buildStepDocuments(
 )
 const index = jsSearchIndex(docs)
 
-configure(monaco, index, expressions)
-
-monaco.editor.create(document.getElementById('root'), {
-  value: `@foo
+const value = `@foo
 Feature: Hello
-  Scenario: Hi
-    Given I have 58 cukes in my belly
-    And this is an undefined step
-`,
+Scenario: Hi
+  Given I have 58 cukes in my belly
+  And this is an undefined step
+    | some | poorly |
+    | formatted | table |
+`
+
+const editor1 = monaco.editor.create(document.getElementById('editor1'), {
+  value,
   language: 'gherkin',
   theme: 'vs-dark',
   // semantic tokens provider is disabled by default
   'semanticHighlighting.enabled': true
-});
+})
+
+const editor2 = monaco.editor.create(document.getElementById('editor2'), {
+  value,
+  language: 'gherkin',
+  theme: 'vs-dark',
+  // semantic tokens provider is disabled by default
+  'semanticHighlighting.enabled': true
+})
+
+const configureEditor = configureMonaco(monaco, index, expressions)
+
+configureEditor(editor1)
+configureEditor(editor2)