+++ react/CHANGELOG.md
@@ -11,15 +11,41 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Changed
 
-* The React components are exported via the `components` property.
-  ([#1660](https://github.com/cucumber/common/pull/1660)
-   [aslakhellesoy])
-
 ### Deprecated
 
 ### Removed
 
 ### Fixed
+
+## [18.1.2] - 2021-09-08
+
+### Fixed
+
+* The 8.1.1 release was accidentally released without compiling the TypeScript code
+
+## [18.1.1] - 2021-09-08
+
+### Fixed
+
+* The 8.1.0 release was accidentally released without compiling the TypeScript code
+
+## [18.1.0] - 2021-09-08
+
+### Added
+
+* Make `<GherkinDocument>` customiseable.
+
+## [18.0.0] - 2021-09-02
+
+### Changed
+
+* The React components are exported via the `components` property.
+  ([#1660](https://github.com/cucumber/common/pull/1660)
+   [aslakhellesoy])
+* Upgrade to `@cucumber/gherkin-utils` `^7.0.0`
+* Upgrade to `@cucumber/messages` `^17.1.0`
+* Upgrade to `@cucumber/query` `^11.0.0`
+* Upgrade to `@cucumber/tag-expressions` `4.0.0`
 
 ## [17.0.0] - 2021-07-08
 
@@ -404,7 +430,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * First release
 
 <!-- Releases -->
-[Unreleased]: https://github.com/cucumber/cucumber/compare/react/v17.0.0...main
+[Unreleased]: https://github.com/cucumber/cucumber/compare/react/v18.1.2...main
+[18.1.2]:      https://github.com/cucumber/cucumber/compare/react/v18.1.1...react/v18.1.2
+[18.1.1]:      https://github.com/cucumber/cucumber/compare/react/v18.1.0...react/v18.1.1
+[18.1.0]:      https://github.com/cucumber/cucumber/compare/react/v18.0.0...react/v18.1.0
+[18.0.0]:      https://github.com/cucumber/cucumber/compare/react/v17.0.0...react/v18.0.0
 [17.0.0]:      https://github.com/cucumber/cucumber/compare/react/v16.2.0...react/v17.0.0
 [16.2.0]:      https://github.com/cucumber/cucumber/compare/react/v16.1.0...react/v16.2.0
 [16.1.0]:      https://github.com/cucumber/cucumber/compare/react/v16.0.2...react/v16.1.0